Defined new stocks list as of Mar 9, updated code for pulling all historic data
A new list of all stocks available in the NYSE has been added, which contains all stocks that are available as of March 9th 2022.
It does not matter if stocks that are no longer available are missing for the dataset as we won't be able to live trade those stocks, so it doens't make sense to use them.
The new stock list is extracted from data downloaded directly from the relevant exchange, so it is most accurate with what stocks are available, and the download does not fail as with yfinance

A new data source will be used (IEX Cloud). In Brainstorm.ipynb there is some code for pulling ALL historic data for a single stock, tested on facebook stock.
Further fields will be added. IEX Cloud has a rate limit, so code to pull all historic data possible of all stocks in NYSE will be added (then from there we can use python to filter for specific dates).
This will allow us to move past any issues of stocks that would not pull, or having to rewrite code for pulling data everytime.