Started work for multistock training
Currently the idea is to have each ticker in a dictionary (fast indexing and pulling cause hashmap), then for each iteration you
create a dataloader, randomly choose a batch from that one stock, use it, purge from memory and move on to the next stop.

parameters for mutistock training will need to be passed to the train function.
Current possible issues that will arise is how to deal with validation and baseline for multiple model