Models hooked, absolute path, change for machine
The code to connect a single stock model and a multistock model is present. However, the path is absolute, so this must be changed so that the path may be relative.

Furthermore, currently the output of the model is percentage change (since we forgoed minmax scaling). Still need to multiply this by the previous close price to get the absolute price prediction.